Move RingBufferQueue to common/
